2.1 Background Colors
Background colors can be thought of as something like an “undercoat” in CSS.
They form the base of your site’s visuals and set the tone for your design. In this lesson, learn how to add background colors and the different types of color codes you can use.
Background-color
Hexcode
RGB